==> R/scripts/03_update_banter_classifier.Rmd <==
---
title: "03_update_classifier"
output: pdf_document
date: "2023-07-31"
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r set up workspace}
## load required packages
library(ggplot2)
library(PAMpal)
library(banter)
library(dplyr)
library(here)

#Set time zone to UTC
Sys.setenv(TZ = 'UTC')

# # install PAMpal package from GitHub
# devtools::install_github('TaikiSan21/PAMpal')
## install BANTER package from GitHub
# devtools::install_github('ericarcher/banter')

```

Update 2013 data with updated PAMpal

```{r run 2013 data through updated PAMpal}
## Create a settings object (loading databases and binaries)
myPps <- PAMpalSettings()
Monodontid_2013_data <- processPgDetections(myPps,mode='db',id='Narluga_043031')
Monodontid_2013_data <- setSpecies(Monodontid_2013_data, method = 'pamguard')
Monodontid_2013_data_ch1 <- filter(Monodontid_2013_data, Channel == '1')

## calculate ICI and add to object
Monodontid_2013_data_ch1 <- calculateICI(Monodontid_2013_data_ch1, time = 'UTC', callType="click")
save(Monodontid_2013_data_ch1, file='Rdata/Monodontid_2013_data_ch1.rdata')

```

Add TOL event level data to AcousticStudy objects

```{r add TOL - training data}
## load data
load(here("R/Rdata/Monodontid_2013_data_ch1.rdata"))

# load csvs with TOL data
train_TOL_df <- read.csv(here('R/spreadsheets/TOL_ratios/training_data_TOL_ratios.csv'))
train_TOL_df_select <- train_TOL_df %>% select(c(eventId,TOL_16to25_sum_dB,TOL_25to40_sum_dB))

# addMeasures function added in PAMpal v0.20.2
# Only requirement for "measures" is an "eventId" column matching event names in our data
train_data_tol <- addMeasures(Monodontid_2013_data_ch1, measures = train_TOL_df_select)
# this will show what we've added for each event
getMeasures(train_data_tol)

# save updated AcousticStudy object with TOL diff
save(train_data_tol, file=here("R/Rdata/train_data_tol.Rdata"))

# these will automatically get exported now with export_banter
# This should tell you that it found 2 event level measures that are exported
train_banterData <- export_banter(train_data_tol)
# should see the band_diff_xxx data with our event data
train_banterData$events

```

```{r add TOL - Fisher data}
## load data - 20 kHz highpass version
load(here("R/Rdata/1hr/Fish_20kHzPG_20kHzPP_1hr_wavTime_filtered_labeled.rdata"))

# load csvs with TOL data
fish_TOL_df <- read.csv(here('R/spreadsheets/TOL_ratios/testing_fish_data_TOL_ratios_0kHz_highpass.csv'))

fish_TOL_df_select <- fish_TOL_df %>% select(c(eventId,TOL_16to25_sum_dB,TOL_25to40_sum_dB))

# addMeasures function added in PAMpal v0.20.2
# Only requirement for "measures" is an "eventId" column matching event names in our data
fish_test_data_tol <- addMeasures(fish_dets_20kHzPP_filter, measures = fish_TOL_df_select)

# this will show what we've added for each event
getMeasures(fish_test_data_tol)

# save updated AcousticStudy object with TOL diff
save(fish_test_data_tol, file=here("R/Rdata/test_data_fish_TOL.Rdata"))
```

```{r add TOL - Kong Oscar data}
## load data
load(here("R/Rdata/1hr/Kong_20kHzPG_20kHzPP_1hr_wavTime_filtered_labeled.rdata"))

# load csvs with TOL data
kong_TOL_df <- read.csv(here('R/spreadsheets/TOL_ratios/testing_kong_data_TOL_ratios_0kHz_highpass.csv'))

kong_TOL_df_select_tol <- kong_TOL_df %>% select(c(eventId,TOL_16to25_sum_dB,TOL_25to40_sum_dB))

# addMeasures function added in PAMpal v0.20.2
# Only requirement for "measures" is an "eventId" column matching event names in our data
kong_test_data_tol <- addMeasures(kong_dets_20kHzPP_filter, measures = kong_TOL_df_select_tol)

# this will show what we've added for each event
getMeasures(kong_test_data_tol)

# save updated AcousticStudy object with TOL diff
save(kong_test_data_tol, file=here("R/Rdata/test_data_kong_TOL.Rdata"))
```

Train banter model with 2013 data

```{r rerun Banter training model - Only Det 2 and 3 - no TOL vars}
## export banter model
load(here("R/Rdata/Monodontid_2013_data_ch1.rdata"))

banterAll <- export_banter(Monodontid_2013_data_ch1,dropVars = c('noiseLevel','dBPP','peakTime'))

## initialize banter model
bant.mdl <- initBanterModel(banterAll$events)

## look at detector names and order
names(banterAll$detectors)

## run RF models for each Detector added-------------------------------------
bant.mdl <- addBanterDetector(
  bant.mdl, 
  data = banterAll$detectors[c(2,3)], # includes only Detectors 2 and 3
  ntree = 10000, 
  sampsize = 50,
  importance = TRUE
)
## look at summary of Detector models
## this shows correct classification rate for each species in each detector
summary(bant.mdl)
plotDetectorTrace(bant.mdl)

## run Event model------------------------------------------------------------
bant.mdl <- runBanterModel(bant.mdl, ntree = 10000, sampsize = 9)

## look at summary for Event model
summary(bant.mdl_TOL)

## get RF data from banter model
event.rf <- getBanterModel(bant.mdl_TOL, "event")

## OTHER THINGS TO LOOK INTO: -------------------------------
## get RF data for specific detector - example of Det 5:
(event.rf.Det4 <- getBanterModel(bant.mdl_TOL, "Click_Detector_4"))

## examine model stability
rfPermute::plotTrace(event.rf)
rfPermute::plotInbag(event.rf)

## get predictor names for event level RF model
colnames(bant.mdl@model.data)

## save model
save(bant.mdl, file=here("R/Rdata/train_bant_mdl.Rdata"))

```

```{r rerun Banter training model - Only Det 2 and 3}
## export banter model
load(here("R/Rdata/train_data.Rdata"))

banterAll <- export_banter(train_data,dropVars = c('noiseLevel','dBPP','peakTime'))

## initialize banter model
bant.mdl <- initBanterModel(banterAll$events)

## look at detector names and order
names(banterAll$detectors)

## run RF models for each Detector added-------------------------------------
bant.mdl <- addBanterDetector(
  bant.mdl, 
  data = banterAll$detectors[c(2,3)], # includes only Detectors 2 and 3
  ntree = 10000, 
  sampsize = 50,
  importance = TRUE
)
## look at summary of Detector models
## this shows correct classification rate for each species in each detector
summary(bant.mdl)
plotDetectorTrace(bant.mdl)

## run Event model------------------------------------------------------------
bant.mdl_TOL <- runBanterModel(bant.mdl, ntree = 10000, sampsize = 9)

## look at summary for Event model
summary(bant.mdl_TOL)

## get RF data from banter model
event.rf <- getBanterModel(bant.mdl_TOL, "event")

## SUMMARIES
rfPermute::plotVotes(event.rf)
rfPermute::plotProximity(event.rf)
rfPermute::casePredictions(event.rf)

## OTHER THINGS TO LOOK INTO: -------------------------------
## get RF data for specific detector - example of Det 5:
(event.rf.Det4 <- getBanterModel(bant.mdl_TOL, "Click_Detector_4"))

## examine model stability
rfPermute::plotTrace(event.rf)
rfPermute::plotInbag(event.rf)

## get predictor names for event level RF model
colnames(bant.mdl_TOL@model.data)

## save model
save(bant.mdl_TOL, file=here("R/Rdata/train_bant_mdl_TOL.Rdata"))

```

```{r rerun Banter training model - Only Det 2 and 3 - with TOL vars}
## export banter model
load(here("R/Rdata/train_data_tol.Rdata"))

banterAll <- export_banter(train_data_tol,dropVars = c('noiseLevel','dBPP','peakTime'))

## initialize banter model
bant.mdl <- initBanterModel(banterAll$events)

## look at detector names and order
names(banterAll$detectors)

## run RF models for each Detector added-------------------------------------
bant.mdl <- addBanterDetector(
  bant.mdl, 
  data = banterAll$detectors[c(2,3)], # includes only Detectors 2 and 3
  ntree = 10000, 
  sampsize = 50,
  importance = TRUE
)
## look at summary of Detector models
## this shows correct classification rate for each species in each detector
summary(bant.mdl)
plotDetectorTrace(bant.mdl)

## run Event model------------------------------------------------------------
bant.mdl_TOL <- runBanterModel(bant.mdl, ntree = 10000, sampsize = 9)

## look at summary for Event model
summary(bant.mdl_TOL)

## get RF data from banter model
event.rf <- getBanterModel(bant.mdl, "event")

## SUMMARIES
rfPermute::plotVotes(event.rf)
rfPermute::plotProximity(event.rf)
rfPermute::casePredictions(event.rf)

## save model
save(bant.mdl_TOL, file=here("R/Rdata/train_bant_mdl_TOL.Rdata"))

```
