==> R/scripts/04_test_banter_classifier.Rmd <==
---
title: "test_beluga_narwhal_classifier"
author: "Marie Zahn"
date: '2022-07-12'
output: pdf_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r set up workspace}
## load required packages
library(ggplot2)
library(PAMpal)
library(banter)
library(dplyr)
library(randomForest)
library(rfPermute)
library(here)

#Set time zone to UTC
Sys.setenv(TZ = 'UTC')

## install PAMpal package from GitHub
# devtools::install_github('TaikiSan21/PAMpal')
## install BANTER package from GitHub
# devtools::install_github('ericarcher/banter')

```

```{r modified code from rfPermute confusionMatrix func}
# function to calculate prediction interval from output of predict() function
confusionMatrix_predict <- function(predict_output) {
  ## make confusion matrix
  cm <- predict_output$validation.matrix
  
  # binomial test
  conf.level = 0.95
  threshold = NULL
  
  # Confusion matrix and counts
  # cm <- .confMat(rf)
  class.n <- rowSums(cm)
  all.n <- c(class.n, Overall = sum(class.n))
  total.n <- sum(class.n)
  correct.n <- diag(cm)
  correct.n <- c(correct.n, Overall = sum(correct.n))
    
  # Confidence intervals
  ci <- t(sapply(
    mapply(
      stats::binom.test, 
      x = correct.n, 
      n = all.n, 
      p = correct.n / all.n, 
      conf.level = conf.level,
      SIMPLIFY = FALSE
    ),
    function(x) x$conf.int * 100
  ))
  colnames(ci) <- paste(c("LCI", "UCI"), conf.level, sep = "_")
  
  # Probability threshold
  prob.gt <- NULL
  if(!is.null(threshold)) {
    prob.gt <- lapply(threshold, function(p) {
      stats::pbinom(correct.n, total.n, p)
    })
    prob.gt <- do.call(cbind, prob.gt)
    colnames(prob.gt) <- paste0("Pr.gt_", threshold)
  }
  
  cm <- rbind(cm, Overall = rep(NA, ncol(cm)))
  pct.correct <- (correct.n / all.n) * 100
  cbind(
    cm, 
    pct.correct = pct.correct[rownames(cm)], 
    ci[rownames(cm), , drop = FALSE],
    prob.gt
  )
}

```

## Test banter model with SoundTrap data - without TOL vars

```{r load data without TOL}
## base model without TOL variables
load(file=here('R/Rdata/train_bant_mdl.rdata'))

## load testing data
## these data were processed in PAMpal with 20 kHz highpass filter and events that had few detections were removed
load(here("R/Rdata/1hr/Fish_20kHzPG_20kHzPP_1hr_wavTime_filtered_labeled.rdata"))
load(here("R/Rdata/1hr/Kong_20kHzPG_20kHzPP_1hr_wavTime_filtered_labeled.rdata"))
```

```{r test whale model without TOL vars}
## remove noise events
fish_whale_data <- fish_dets_20kHzPP_filter %>% filter(species!='NOISE')
kong_whale_data <- kong_dets_20kHzPP_filter %>% filter(species!='NOISE')

## Export data from AcousticStudy into format required to run a BANTER model
## Fisher islands
banter_fish <- export_banter(fish_whale_data,
                             dropVars = c('noiseLevel', 'dBPP', 'peakTime'))
## Kong Oscar
banter_kong <- export_banter(kong_whale_data,
                             dropVars = c('noiseLevel', 'dBPP', 'peakTime'))

# filter out clicks that were not detected in click detector 2
banter_fish$detectors$Click_Detector_2<-filter(banter_fish$detectors$Click_Detector_2, peak>20 & peak<50)
banter_fish$detectors$Click_Detector_3<-filter(banter_fish$detectors$Click_Detector_3, peak>50 & peak<70)

banter_kong$detectors$Click_Detector_2<-filter(banter_kong$detectors$Click_Detector_2, peak>20 & peak<50)
banter_kong$detectors$Click_Detector_3<-filter(banter_kong$detectors$Click_Detector_3, peak>50 & peak<70)

# predict species with new data
score_fish_whale <- predict(bant.mdl, banter_fish)
confusionMatrix_predict(score_fish_whale)

score_kong_whale <- predict(bant.mdl, banter_kong)
confusionMatrix_predict(score_kong_whale)

## look at beluga predictions
score_fish_whale$predict.df[score_fish_whale$predict.df$original == "X045",]
score_kong_whale$predict.df[score_kong_whale$predict.df$original == "X045",]

## look at narwhal predictions
score_fish_whale$predict.df[score_fish_whale$predict.df$original == "X085",]
score_kong_whale$predict.df[score_kong_whale$predict.df$original == "X085",]

```

```{r save classification scores}
save(score_fish_whale,file=here('R/Rdata/score_fish_whale.rdata'))
save(score_kong_whale,file=here('R/Rdata/score_kong_whale.rdata'))

```

## Test banter model with SoundTrap data - with TOL vars

```{r load data}
## load banter model with TOL variables from specific freq bands
load(file=here('R/Rdata/train_bant_mdl_TOL.rdata'))

## load testing data that includes TOL diff calculated with 0 kHz highpass filter
load(here("R/Rdata/test_data_fish_TOL.Rdata"))
load(here("R/Rdata/test_data_kong_TOL.Rdata"))

```

```{r test whale model with TOL vars}
## remove noise events
fish_whale_data <- fish_test_data_tol %>% filter(species!='NOISE')
kong_whale_data <- kong_test_data_tol %>% filter(species!='NOISE')

## Export data from AcousticStudy into format required to run a BANTER model
## Fisher islands
banter_fish_TOL <- export_banter(fish_whale_data,
                             dropVars = c('noiseLevel', 'dBPP', 'peakTime'))
## Kong Oscar
banter_kong_TOL <- export_banter(kong_whale_data,
                             dropVars = c('noiseLevel', 'dBPP', 'peakTime'))

# filter out clicks that were not detected in click detector 2
banter_fish_TOL$detectors$Click_Detector_2<-filter(banter_fish_TOL$detectors$Click_Detector_2, peak>20 & peak<50)
banter_fish_TOL$detectors$Click_Detector_3<-filter(banter_fish_TOL$detectors$Click_Detector_3, peak>50 & peak<70)

banter_kong_TOL$detectors$Click_Detector_2<-filter(banter_kong_TOL$detectors$Click_Detector_2, peak>20 & peak<50)
banter_kong_TOL$detectors$Click_Detector_3<-filter(banter_kong_TOL$detectors$Click_Detector_3, peak>50 & peak<70)

# predict species with novel data
# score_Fisher <- predict(bant.mdl_new, banter_Fisher)
score_fish_whale_TOL <- predict(bant.mdl_TOL, banter_fish_TOL)
confusionMatrix_predict(score_fish_whale_TOL)

score_kong_whale_TOL <- predict(bant.mdl_TOL, banter_kong_TOL)
confusionMatrix_predict(score_kong_whale_TOL)
```

# Assess variable importance and percent votes

```{r plotAssignments from rfPermute}
# forked code from Eric Archer's github swfscMisc
plotAssignments <- function(
  probs, orig, type = NULL, ylab = NULL, freq.sep.line = TRUE, plot = TRUE
) {
  freq <- table(orig)
  type <- if(is.null(type)) {
    if(all(freq <= 30)) "bar" else "area"
  } else {
    match.arg(type, c("bar", "area"))
  }
  
  if(is.null(colnames(probs))) {
    colnames(probs) <- paste("Group", 1:ncol(probs), sep = ".")
  }
  
  df <- data.frame(orig = orig, probs, check.names = FALSE)
  i <- do.call(order, c(as.list(df), list(decreasing = TRUE)))
  df <- df[i, ] %>% 
    dplyr::mutate(id = 1:dplyr::n()) %>% 
    tidyr::pivot_longer(-c("id", "orig"), names_to = "pred", values_to = "prob")
  
  df$orig <- as.factor(df$orig)
  levels(df$orig) <- c("Beluga","Narwhal")
  names(freq) <- c("Beluga","Narwhal")
  
  if(!is.null(freq.sep.line)) {
    levels(df$orig) <- paste0(
      names(freq), ifelse(freq.sep.line, "\n", " "), "(n = ", freq, ")"
    )
  }
  
  if(is.null(ylab)) ylab <- "Assignment Probability"
  
  p <- ggplot2::ggplot(df, ggplot2::aes_string("id", "prob")) +
    switch(
      type,
      area = ggplot2::geom_area(
        ggplot2::aes_string(fill = "pred"), 
        stat = "identity"
      ),
      bar = ggplot2::geom_bar(
        ggplot2::aes_string(fill = "pred"), 
        stat = "identity"
      )
    ) +
    ggplot2::scale_fill_discrete(
      guide = ggplot2::guide_legend(title = "Predicted")
    ) +
    ggplot2::scale_x_continuous(expand = c(0, 0)) +
    ggplot2::scale_y_continuous(expand = c(0, 0)) +
    ggplot2::facet_wrap(~ orig, scales = "free_x") +
    ggplot2::ylab(ylab) +
    ggplot2::theme(
      legend.position = "top",
      text = ggplot2::element_text(size = 14),
      axis.text.x = ggplot2::element_blank(),
      axis.title.x = ggplot2::element_blank(),
      axis.ticks.x = ggplot2::element_blank(),
      plot.title = ggplot2::element_text(hjust = 0),
      panel.grid = ggplot2::element_blank(),
      panel.background = ggplot2::element_blank()
    )
  if(plot) print(p)
  invisible(p)
}
```


```{r plotVotes}
# plotVotes of model predictions
# base model without TOL measures
# BANTER predicts fisher
plotAssignments((score_fish_whale$predict.df[,3:4]/rowSums(score_fish_whale$predict.df[,3:4]))*100,
                score_fish_whale$predict.df$original,ylab = "Percent of votes")

# BANTER predicts kong oscar
plotAssignments((score_kong_whale$predict.df[,3:4]/rowSums(score_kong_whale$predict.df[,3:4]))*100,
                score_kong_whale$predict.df$original,ylab = "Percent of votes")

# BANTER models with TOL measures for supplementary figure
# BANTER predicts fisher
plotAssignments((score_fish_whale_TOL$predict.df[,3:4]/rowSums(score_fish_whale_TOL$predict.df[,3:4]))*100,
                score_fish_whale_TOL$predict.df$original,ylab = "Percent of votes")

ggsave(here("R/figures/plotVotes_banter-predicts-fisher.png"), dpi=300)

# BANTER predicts kong oscar
plotAssignments((score_kong_whale_TOL$predict.df[,3:4]/rowSums(score_kong_whale_TOL$predict.df[,3:4]))*100,
                score_kong_whale_TOL$predict.df$original,ylab = "Percent of votes")

ggsave(here("R/figures/plotVotes_banter-predicts-kong.png"), dpi=300)

```
